+++ include/build.h
@@ -2,6 +2,6 @@
  * Author: sascha_lammers@gmx.de
  */
 #pragma once
-#define __BUILD_NUMBER "13155"
-#define __BUILD_NUMBER_INT 13155
+#define __BUILD_NUMBER "13160"
+#define __BUILD_NUMBER_INT 13160