+++ scripts/tools/kfcfw_discovery.py
@@ -1,7 +1,7 @@
 from zeroconf import ServiceBrowser, Zeroconf
 
 TYPE = "_kfcmdns._udp.local."
-FORMAT = "%-16.16s | %-22.22s | %s | %-16.16s | %-24.24s | %-19.19s | %-12.12s"
+FORMAT = "%-16.16s | %-22.22s | %s | %-16.16s | %-32.32s | %-19.19s | %-12.12s"
 
 class MyListener: