+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.0.8 (master)
 
+ - Support for INMP441 microphone for the audio visualizer
  - Changed flash frequency and mode to 80MHz/QIO for ESP8266
  - Changed nvs_flash to default config storage for ESP8266
  - Ported nvs_flash from ESP8266_RTOS_SDK to ESP8266_NONOS_SDK ([https://github.com/sascha432/ESP8266_NvsFlash_NONOS_SDK](https://github.com/sascha432/ESP8266_NvsFlash_NONOS_SDK))